api/requirements.txt
pandas>=2.0
numpy>=2.3.4
Faker>=23.0
pydantic>=2.11

psycopg2-binary>=2.9
python-dotenv>=1.0
SQLAlchemy>=2.0

fastapi>=0.105
uvicorn>=0.25

python-multipart>=0.0.6